feat(proj): ticket-1823 add type-enum to package.json
add type-enum rule that defines the types we could use when creating a commit message, ie: major,
minor, patch, wip.
If a new type has been added to the rule, you have to edit the .cz-config.js in
order to see the new types in the predefined commit messages, and for the commit message to be
pushed successfully.
Other configurations can be edited in  the .cz-config.js file

ISSUES CLOSED: 1823